+++ R/package-maotai.R
@@ -20,7 +20,7 @@
 #' @import Rdpack
 #' @importFrom stats rnorm
 #' @importFrom shapes procGPA
-#' @importFrom mclust Mclust
+#' @importFrom mclust Mclust mclustBIC
 #' @importFrom utils packageVersion
 #' @importFrom RSpectra eigs
 #' @importFrom Matrix rankMatrix